Refactored codebase to fix issues with user authentication
Updated NavUserDropdown component to check for the presence of an access token in local storage and set a loggedIn state to true if present.

Modified the prepareHeaders function in the apiSlice to set the authorization header with the access token in the Redux state.

Added a forgotPassword mutation to the extendedUserSlice.

Fixed a bug in the ForgotPassword component where the forgotPassword mutation was not being called on form submission.

Modified the Login component to fix a bug where the password input was not being masked.

Fixed a bug in the Register component where an error was not being displayed when passwords did not match.

Other minor fixes and improvements to improve overall code quality and functionality.