Create [accessToken].tsx
Add "Reset Password" component with a form that has an email input field. When the form is submitted, the server should send an email to the user containing a link to reset their password. After the user enters a new password, the component should update the password in the database and redirect the user to the login page.